requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.1
attrs==21.4.0
av==9.1.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.4b2
bleach==4.1.0
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chainer==7.8.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@42c44b622eb7b179576bdac218c84c107fff8eab
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fastapi==0.72.0
filelock==3.6.0
filetype==1.0.9
flatbuffers==1.12
fonttools==4.31.2
freetype-py==2.2.0
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h11==0.13.0
h5py @ file:///tmp/build/80754af9/h5py_1637138488546/work
huggingface-hub==0.5.1
hydra-core==1.1.2
icecream==2.1.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
intrinsics==0.0.0
iopath==0.1.9
ipdb==0.13.9
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
kornia==0.6.4
libclang==14.0.1
llvmlite==0.36.0
lmdb==1.3.0
lpips==0.1.4
lpips-pytorch===latest
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpld3==0.5.8
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
nose==1.3.7
notebook==6.4.10
numba==0.53.1
numexpr @ file:///tmp/build/80754af9/numexpr_1640689833592/work
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
onnxruntime==1.10.0
onnxruntime-gpu==1.10.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
piq==0.7.0
plotly==5.7.0
portalocker==2.4.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.9.0
pyDeprecate==0.3.1
pydot==1.4.2
pyglet==1.5.23
Pygments==2.11.2
PyMatting==1.1.5
PyMCubes==0.1.2
PyOpenGL==3.1.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
pyrender==0.1.45
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-multipart==0.0.5
pytorch-fid==0.2.1
pytorch-lightning==1.5.10
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
rawpy==0.17.1
regex==2022.3.15
rembg==2.0.21
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
scikit-image==0.19.1
scikit-learn==1.0.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
starlette==0.17.1
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.3.25
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
torch-fidelity==0.3.0
torchmetrics==0.7.3
torchvision==0.12.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.18.0
trimesh==3.10.5
tru==0.0.0
tu==0.0.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
uvicorn==0.17.0
vit-pytorch==0.32.2
wandb==0.12.14
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.1
yacs==0.1.8
yarg==0.1.9
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
